Supermarket sales Analysis
The data consist of 3 months transaction details of a supermarket with 3 branches in 3 different cities. Where data has 17 variables with 1000 transactions. The supermarket has two types of customers Member & Normal. There are 6 different Product categories are sold in these supermarkets. E-wallet, cash, credit card payments are accepted in these supermarkets. The Supermarkets are open all days from Monday- to Sunday. Timings of a supermarket are from 10 am to 8 pm. All transactions in the supermarket have a unique Invoice ID.