Super Market Analysis (Market basket Analysis)
Data: Data consists of 6 columns in the data with equal size. All the columns in the data are in object type and have the same number of rows i.e 372610. Order ID is customer permanent ID for the transactions in-store, we can identify with the number of unique values that few Order IDs have repeated transactions. Product: We have 20 different electronic products which are given in the above result. Quantity: No of products purchased in a transaction. Price: Price Of each product purchased in a transaction. Order date: Date of order for a transaction. Purchase Address: Customer address where the product has to be delivered.

Null Values All the columns contain an equal number of null values.

After basic data cleaning: data reduced to 185686 rows with 6 columns and we have converted columns to respective data type,dropped the values with same repeating rows(duplicate values). Excluded the columns which are not required , removed null values in the data, and renamed the columns for our convince.

Created new columns splitting address & date. Analyzed graphs on the basis of questions.

Conclusion of Analysis: According to the observation of the plots, San Francisco can be the best city for establishing a new branch for the retail store because it has the highest sales. And most sold products are AAA batteries of 4 pack. High sales are happening in the evening from Morning 8 am to 4 pm, the sales are good enough. December month has highest sales . Through Market basket analysis we found out together frequently bought items.

​