HR Analytics
The objective is to use the employee's data to know the factors affecting the salaries of an employee. And build a model that can predict the salaries varying with respect to considerable factors and interpret the results to find the factors which influence the score. Data has 5000 rows and 20 features. Data contains an object, integer, float type data type where the file is in CSV format.

There are freshers and up to 6 years of experienced employees in the data.

The data includes the Salaries of employees, which makes this a supervised regression machine learning task:

Supervised: we have access to both the features and the target and our goal is to train a model that can learn a mapping between the two Regression: The Salary is the target variable which is dependent. All other variables are Independent variables.

Approach:

Data exploration and data visualization for exploring each variable w.r.t salaries for both categorical and numerical variables to know the importance of each variable.

Cleaning data like handling missing values, outliers, data types, etc.

With the below variables, we can know if Gender, Age, Base pay, or any of the variables is affecting salary in a positive or a negative way with various techniques like correlation, t-test, Multicollinearity.

And then bring all the variables to equal scale using feature scaling. Then split the data into train and test. Where the train is to train the model and test is to test the model.

Build a regression model which is best suitable for our data after analyzing data.

Then use an evaluation metric for the regression model adjusted R square which will evaluate our model accuracy and the importance of variables.